chore: Update Scénario class with package references and toString methods for City, Department, Station, and Year classes

+++ Dev Orienté Obj/src/modele/Year.java
@@ -92,12 +92,12 @@ public class Year {
      * 
      * @return The average price of the year.
      */
-    public double getAverageRate() {
+    public float getAverageRate() {
         if (averagePricePerCity == null || averagePricePerCity.isEmpty()) {
             throw new IllegalArgumentException("No average price data available for this year");
         }
 
-        double sum = 0;
+        float sum = 0;
         for (Float price : averagePricePerCity.values()) {
             sum += price;
         }
